0428 temp version

+++ train_src/loss_func.py
@@ -66,7 +66,7 @@ class DiceBCELoss(nn.Module):
         
         intersection = (inputs * targets).sum()                            
         dice_loss = 1 - (2.*intersection + smooth)/(inputs.sum() + targets.sum() + smooth)  
-        BCE = F.binary_cross_entropy(inputs, targets, reduction='mean')
+        BCE = F.binary_cross_entropy(inputs.float(), targets.float(), reduction='mean')
         Dice_BCE = BCE + dice_loss
         
         return Dice_BCE
